requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.3.0
ccxt>=3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dotenv==0.9.9
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
propcache==0.3.1
pycares==4.6.0
pycparser==2.22
python-dotenv>=1.0.0
PyYAML>=6.0
requests==2.32.3
setuptools==78.1.0
sortedcontainers>=2.4.0
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.19.0
